requirements.txt
requests
pytest
pytest-html
PyYAML
jinja2
faker
pyyaml